Add separate project directories.
