+++ patterns/memento/memento_editor/application.dart
@@ -1,4 +1,5 @@
 import 'editor/editor.dart';
+import 'memento_pattern/originator.dart';
 
 class MementoEditorApplication {
   final editor = Editor();
