+++ patterns/memento/memento_editor/editor/manipulator.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
 import '../../../adapter/flutter_adapter/classic_app/classic_app.dart';
-import 'shapes.dart';
+import '../shapes/shapes.dart';
 
 mixin Manipulator implements ClassicApp, Shapes {
   void paintSelectFrame(Canvas canvas) {
