+++ patterns/memento/memento_editor/shapes/shape.dart
